Quiz App
This is an interactive Quiz App built using HTML, CSS, and JavaScript. It provides users with a set of quiz questions, a timer, and an intuitive interface to choose answers. The app tracks and displays the user's score, the correct and incorrect answers, and provides a recap at the end. It includes features like randomized answer choices, a countdown timer, and a restart option. This project showcases my skills in front-end development, including JavaScript functions, DOM manipulation, event handling, and dynamic content updating.