Minion Eyes
The "Minion Eyes" project is a fun and interactive web design concept with a playful animated design. The main highlight of this project is a pair of animated eyes that track the user's mouse movement, mimicking the playful nature of Minion characters. Built with HTML, CSS, and JavaScript, this project incorporates Bootstrap for responsive design, ensuring an excellent user experience across all devices. The page includes a header, navigation bar, home, about, contact information sections, and a dynamic contact form. With its modern styling and eye-catching animations, this project showcases a mix of creativity and technical skills, perfect for anyone interested in web development and UI/UX design.