Image Resize
This project is an Image Resizer and Compressor web application that allows users to easily resize and compress images. The app provides a user-friendly interface where users can upload an image, adjust its dimensions by pixel or percentage, and control the image quality using a range slider. Once the image is resized and compressed, users can download the optimized image. The app uses HTML, CSS, and JavaScript, along with the power of HTML5 Canvas for image manipulation. It also offers real-time image previews and dynamic resizing options, making it perfect for anyone looking to quickly reduce the size of an image without compromising much on quality.