Guess the Anime
I’ve built an interactive Guess the Anime game using HTML, CSS, and JavaScript! This web app challenges users to guess the name of an anime from a series of emoji-based hints. The game includes dynamic letter buttons, a countdown timer for tries, and a responsive design for a seamless device experience. Combining UI/UX principles with JavaScript logic to create an engaging user interface was a fun challenge. It's a great way to practice web development and front-end design skills!